shim: split settings+util
